pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dev-kit"
version = "0.1.0"
description = "Kit personal de desarrollo: scripts ETL, templates y herramientas de datos"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "openpyxl>=3.1",
    "pyarrow>=16.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pandas-stubs>=2.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

# ─── Black ────────────────────────────────────────────────────
[tool.black]
line-length = 88
target-version = ["py39"]

# ─── isort ────────────────────────────────────────────────────
[tool.isort]
profile = "black"
line_length = 88

# ─── Ruff ─────────────────────────────────────────────────────
[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

# ─── Mypy ─────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true

# ─── Pytest ───────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
